requirements.txt
essentia>=2.1b6.dev1110
httpx>=0.27.0
beautifulsoup4>=4.12.0
google-generativeai>=0.8.0
pydantic>=2.4.0
pydantic-settings>=2.0.0
numpy>=1.24.0
